+++ LIBRARY/src/bid64_scalb.c
@@ -51,14 +51,14 @@ BID_TYPE0_FUNCTION_ARGTYPE1_OTHER_ARGTYPE2(BID_UINT64, bid64_scalbn, BID_UINT64,
        exp64 = (BID_SINT64) exponent_x + (BID_SINT64) n;
        if(exp64<0) exp64=0;
        if(exp64>MAX_DECIMAL_EXPONENT) exp64=MAX_DECIMAL_EXPONENT;
-       exponent_x = exp64;
+       exponent_x = (int) exp64;
       res = very_fast_get_BID64 (sign_x, exponent_x, coefficient_x);	// 0
     }
     BID_RETURN (res);
   }
 
   exp64 = (BID_SINT64) exponent_x + (BID_SINT64) n;
-  exponent_x = exp64;
+  exponent_x = (int) exp64;
 
   if ((BID_UINT32) exponent_x <= MAX_DECIMAL_EXPONENT) {
     res = very_fast_get_BID64 (sign_x, exponent_x, coefficient_x);
